Update get_current_user dependency to use new token use case. Also update the database from an in-memory database to a file database (explained below). Also create new endpoint /users/me
Why the in-memory database has been changed:
This has been the source of a lot of pain. While implementing the /users/me endpoint (with the new dependency) I was running into this error: no such table 'tokens'. This was baffling; the table did exist - in fact the token which was being used as an authentication header for that request had just been inserted into the table. How could it be that the table didn't exist? After lots of debugging (and we're talking about 2 hours) to ensure that the code I had written was indeed sound (and frustratingly, it was), I arrived at a point in the debugger: I could run queries when stepping through the token *generation* code (i.e. using the session in that use case) but not in the token *verification* code (i.e. the session in *that* use case). There was a problem with the way that the sessions were being generated. Once again, I was baffled; they were using the same get_token_repository() function which created the same session objects bound to the same engine (which had all the tables created in it). Having already wasted an evening on this, I resigned myself to bed, defeated. The following morning's shower summoned a wonderful lightbulb moment. There *was* indeed a difference in the way that the sessions were being created. The token *validation* code was being called in a FastAPI Dependency. Notably, this dependency is NOT async. A now-not-so-useless piece of information I learned while reading the FastAPI docs is that if a function is declared as not async (i.e. you create it as a normal def) and it is used as either a path operation function, OR a dependency, FastAPI does some magic with asyncio and runs the function on another thread (to not be blocking). Another piece of information is that SQLite in-memory databases are per-thread! You can see where this is going... The verification code was being ran in a synchronous dependency on another "thread" (or whatever you should call threads in GIL land), thus, in that "thread"'s version of the in-memory database, the tokens table really did not exist.

Using a file database fixes this issue, because now all the data (and the schema) gets shared between "threads", and the dependencies see the same database, and things work. What a wonderful bug...